Code refactoring to match the new files
